shallot: fold gpu-particles into its recipe
The shallot-gpu-particles workspace package is gone. Its plugin lives in
examples/recipes/gpu-particles/src/particles and the recipe's shallot.json
enables it by path, so a copied-out recipe installs with no unpublished
dependency.
