+++ examples/recipes/gpu-particles/package.json
@@ -5,7 +5,6 @@
     "type": "module",
     "dependencies": {
         "@dylanebert/shallot": "file:../../..",
-        "shallot-gpu-particles": "workspace:*",
         "typegpu": "~0.12.4"
     }
 }
