+++ examples/recipes/gpu-particles/src/particles/kernel.ts
