+++ examples/recipes/gpu-particles/src/particles/particles.ts
