--- packages/shallot-gpu-particles/package.json
@@ -1,14 +1,0 @@
-{
-    "name": "shallot-gpu-particles",
-    "version": "0.0.0",
-    "private": true,
-    "type": "module",
-    "exports": {
-        ".": "./src/index.ts"
-    },
-    "files": ["src"],
-    "peerDependencies": {
-        "@dylanebert/shallot": "file:../..",
-        "typegpu": "~0.12.4"
-    }
-}
